[DS-2378] Pass 2:  fold SpringServiceManager into DSpaceServiceManager and simplify.
