+++ dspace-services/src/main/java/org/dspace/servicemanager/package-info.java
@@ -7,7 +7,8 @@
  */
 
 /**
- * Register, configure, manipulate services.
+ * Register, configure, manipulate, locate services.
+ * See {@link DSpaceServiceManager} for a starting point.
  */
 
 package org.dspace.servicemanager;
