+++ dspace-services/src/main/java/org/dspace/servicemanager/spring/package-info.java
@@ -9,7 +9,7 @@
 /**
  * Spring extensions and adaptors for DSpace.
  *
- * @see org.dspace.utils.DSpace for starting the kernel.
+ * See {@link org.dspace.utils.DSpace} for starting the kernel.
  */
 
 package org.dspace.servicemanager.spring;
